+++ discord/enums.py
@@ -45,6 +45,7 @@ __all__ = (
     "ActivityType",
     "NotificationLevel",
     "TeamMembershipState",
+    "TeamRole",
     "WebhookType",
     "ExpireBehaviour",
     "ExpireBehavior",
@@ -630,6 +631,15 @@ class TeamMembershipState(Enum):
 
     invited = 1
     accepted = 2
+
+
+class TeamRole(Enum):
+    """Role of a team member."""
+
+    owner = "owner"
+    admin = "admin"
+    developer = "developer"
+    read_only = "read_only"
 
 
 class WebhookType(Enum):