+++ CHANGELOG.md
@@ -12,7 +12,12 @@ These changes are available on the `master` branch, but have not yet been releas
 
 ### Added
 
+- Support for **Python 3.14**.
+  ([#2948](https://github.com/Pycord-Development/pycord/pull/2948))
+
 ### Changed
+
+- Changed that `read_only` team members aren't considered as owner.
 
 ### Fixed
 
@@ -24,6 +29,9 @@ These changes are available on the `master` branch, but have not yet been releas
   ([#3181](https://github.com/Pycord-Development/pycord/pull/3181))
 - Fixed internal use of deprecated \_PayloadLike dict operations.
   ([#3189](https://github.com/Pycord-Development/pycord/pull/3189))
+- Fixed incorrect type hints for `MessagePinIterator`.
+  ([#3178](https://github.com/Pycord-Development/pycord/pull/3178))
+- Fixed team `permissions` keyerror by replacing deprecated permissions with `role`.
 
 ### Deprecated